+++ sdmx-dl-cli/src/main/java/sdmxdl/cli/SearchDatabasesCommand.java
@@ -4,9 +4,8 @@ import internal.sdmxdl.cli.WebSourceOptions;
 import internal.sdmxdl.cli.ext.CsvTable;
 import internal.sdmxdl.cli.ext.RFC4180OutputOptions;
 import picocli.CommandLine;
-import sdmxdl.Flow;
+import sdmxdl.Database;
 import sdmxdl.format.Search;
-import sdmxdl.format.csv.SdmxCsvFields;
 
 import java.io.IOException;
 import java.util.Collection;
@@ -19,8 +18,8 @@ import static java.util.Locale.ROOT;
 /**
  * @author Philippe Charles
  */
-@CommandLine.Command(name = "flows")
-public final class SearchFlowsCommand implements Callable<Void> {
+@CommandLine.Command(name = "databases")
+public final class SearchDatabasesCommand implements Callable<Void> {
 
     @CommandLine.Mixin
     private WebSourceOptions web;
@@ -48,21 +47,22 @@ public final class SearchFlowsCommand implements Callable<Void> {
         return null;
     }
 
-    @SuppressWarnings("unchecked")
-    private CsvTable<Search.Result<Flow>> getTable() {
+    private CsvTable<Search.Result<Database>> getTable() {
         return CsvTable
-                .<Search.Result<Flow>>builder()
-                .columnOf("Ref", result -> result.getItem().getRef(), SdmxCsvFields.getDataflowRefFormatter())
+                .<Search.Result<Database>>builder()
+                .columnOf("Ref", result -> result.getItem().getRef().toString())
                 .columnOf("Name", result -> result.getItem().getName())
-                .columnOf("Description", result -> result.getItem().getDescription())
                 .columnOf("Score", result -> String.format(ROOT, "%.6f", result.getScore()))
                 .build();
     }
 
-    private Stream<Search.Result<Flow>> getRows() throws IOException {
-        Collection<Flow> flows = web.loadManager().usingName(web.getSource()).getFlows(web.toDatabaseRequest());
-        Search<Flow> search = Search.ofFlows(flows);
-        List<Search.Result<Flow>> results = search.search(query, maxResults);
+    private Stream<Search.Result<Database>> getRows() throws IOException {
+        Collection<Database> databases = web.loadManager()
+                .usingName(web.getSource())
+                .getDatabases(web.toSourceRequest());
+        Search<Database> search = Search.ofDatabases(databases);
+        List<Search.Result<Database>> results = search.search(query, maxResults);
         return results.stream();
     }
 }
+